Settings repo and necessary refactoring
