+++ src/LitePatch/LitePatch.Services/Interfaces/ISettingsRepository.cs
@@ -1,0 +1,12 @@
+﻿using LitePatch.Services.Models;
+
+namespace LitePatch.Services.Interfaces;
+
+public interface ISettingsRepository
+{
+    public ApplicationSetting ReadSettings();
+
+    public void WriteSettings(ApplicationSetting settings);
+
+    public void CheckForPath();
+}
