+++ src/LitePatch/LitePatch.Services/Models/ApplicationSetting.cs
@@ -5,4 +5,5 @@ public class ApplicationSetting
     public string RepositoryPath { get; set; } = string.Empty;
 
     public string OutputFolderPath { get; set; } = string.Empty;
+    
 }
