+++ src/LitePatch/LitePatch.Services/Services/SettingsService.cs
@@ -2,59 +2,24 @@
 using System.Text.Json.Serialization;
 using LitePatch.Services.Interfaces;
 using LitePatch.Services.Models;
+using LitePatch.Services.Repo;
 
 namespace LitePatch.Services.Services;
 
 public class SettingsService : ISettingsService
 {
+    public SettingsRepository SettingsRepository { get; set; }
     public ApplicationSetting Settings { get; set; }
 
     public SettingsService()
     {
-        Settings = ReadSettings();
+        SettingsRepository = new SettingsRepository();
+        Settings = SettingsRepository.ReadSettings();
     }
 
     public void Save()
     {
-        WriteSettings();
-    }
-
-    private readonly string _settingsPath = Path.Combine(
-        Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData), 
-        "LitePatch");
-
-    private const string SettingsFile = "settings.json";
-
-    private string SettingsFilePath => Path.Combine(_settingsPath, SettingsFile);
-
-    private void CheckForPath()
-    {
-        if (!Directory.Exists(_settingsPath))
-            Directory.CreateDirectory(_settingsPath);
+        SettingsRepository.WriteSettings(Settings);
     }
     
-    private ApplicationSetting ReadSettings()
-    {
-        CheckForPath();
-
-        if (!File.Exists(SettingsFilePath))
-        {
-            return new ApplicationSetting();
-        }
-
-        var text = File.ReadAllText(SettingsFilePath);
-        return JsonSerializer.Deserialize<ApplicationSetting>(text)!;
-    }
-
-    private void WriteSettings()
-    {
-        CheckForPath();
-
-        var json = JsonSerializer.Serialize(Settings, new JsonSerializerOptions()
-        {
-            WriteIndented = true
-        });
-        
-        File.WriteAllText(SettingsFilePath, json);
-    }
 }
